+++ core/src/main/kotlin/com/abysl/gdxcrawler/Game.kt
@@ -1,6 +1,8 @@
 package com.abysl.gdxcrawler
 
 import com.abysl.gdxcrawler.physics.IPhysics
+import com.abysl.gdxcrawler.screens.game.GameScreen
+import com.abysl.gdxcrawler.screens.main.MainMenu
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.Screen
 import ktx.app.KtxGame
