--- core/src/main/kotlin/com/abysl/gdxcrawler/MainMenu.kt
@@ -1,31 +1,0 @@
-package com.abysl.gdxcrawler
-
-import com.badlogic.gdx.Screen
-import com.badlogic.gdx.graphics.g2d.Sprite
-import com.badlogic.gdx.graphics.g2d.TextureAtlas
-
-class MainMenu : Screen {
-    private val texture = TextureAtlas()
-    private val sprite = Sprite()
-
-    override fun hide() {
-    }
-
-    override fun show() {
-    }
-
-    override fun render(delta: Float) {
-    }
-
-    override fun pause() {
-    }
-
-    override fun resume() {
-    }
-
-    override fun resize(width: Int, height: Int) {
-    }
-
-    override fun dispose() {
-    }
-}
