+++ core/src/main/kotlin/com/abysl/gdxcrawler/ecs/events/EventManager.kt
@@ -4,18 +4,15 @@ import java.util.*
 
 class EventManager {
 
-    // has to be public because of inline fun, would rather be private
     // Have to do a MutableList<Any>, then do unsafe cast later. Couldn't think of another way to do it
-    val subscriptions: MutableMap<Class<out Event>, MutableList<Any>> = mutableMapOf()
+    @PublishedApi
+    internal val subscriptions: MutableMap<Class<out Event>, MutableList<Any>> = mutableMapOf()
     private val events: Queue<Event> = LinkedList()
 
     // don't really understand inline and reified, intellij just told me I need it, otherwise error
     inline fun <reified T : Event> subscribe(noinline subscriber: (T) -> Unit) {
         val event = T::class.java
-        if (subscriptions[event] == null) {
-            subscriptions[event] = mutableListOf()
-        }
-        subscriptions[event]!!.add(subscriber)
+        subscriptions.getOrPut(event) { mutableListOf() }.add(subscriber)
     }
 
     fun publish(event: Event) {
@@ -24,7 +21,7 @@ class EventManager {
 
     private fun <T : Event> processEvent(event: T) {
         subscriptions[event.javaClass]?.forEach {
-            (it as ((T) -> Unit)).invoke(event)
+            (it as (T) -> Unit).invoke(event)
         }
     }
