+++ core/src/main/kotlin/com/abysl/gdxcrawler/screens/game/GameInput.kt
@@ -1,4 +1,4 @@
-package com.abysl.gdxcrawler
+package com.abysl.gdxcrawler.screens.game
 
 import com.abysl.gdxcrawler.ecs.events.EventManager
 import com.abysl.gdxcrawler.settings.InputSettings
@@ -23,26 +23,26 @@ class GameInput(val eventManager: EventManager, var settings: InputSettings) : I
     }
 
     override fun keyTyped(character: Char): Boolean {
-        TODO("Not yet implemented")
+        return false
     }
 
     override fun touchDown(screenX: Int, screenY: Int, pointer: Int, button: Int): Boolean {
-        TODO("Not yet implemented")
+        return false
     }
 
     override fun touchUp(screenX: Int, screenY: Int, pointer: Int, button: Int): Boolean {
-        TODO("Not yet implemented")
+        return false
     }
 
     override fun touchDragged(screenX: Int, screenY: Int, pointer: Int): Boolean {
-        TODO("Not yet implemented")
+        return false
     }
 
     override fun mouseMoved(screenX: Int, screenY: Int): Boolean {
-        TODO("Not yet implemented")
+        return false
     }
 
     override fun scrolled(amount: Int): Boolean {
-        TODO("Not yet implemented")
+        return false
     }
 }
