+++ core/src/main/kotlin/com/abysl/gdxcrawler/screens/game/GameScreen.kt
@@ -1,4 +1,4 @@
-package com.abysl.gdxcrawler
+package com.abysl.gdxcrawler.screens.game
 
 import com.abysl.gdxcrawler.ecs.components.CMove
 import com.abysl.gdxcrawler.ecs.components.CPosition
@@ -6,6 +6,7 @@ import com.abysl.gdxcrawler.ecs.events.EventManager
 import com.abysl.gdxcrawler.ecs.getPlayerArchetype
 import com.abysl.gdxcrawler.physics.IPhysics
 import com.abysl.gdxcrawler.rendering.GameRenderer
+import com.abysl.gdxcrawler.settings.InputSettings
 import com.abysl.gdxcrawler.settings.RenderSettings
 import com.abysl.gdxcrawler.utils.GameWorld
 import com.abysl.gdxcrawler.world.TileWorld
@@ -14,23 +15,31 @@ import com.artemis.ArchetypeBuilder
 import com.artemis.Entity
 import com.artemis.World
 import com.artemis.managers.TagManager
+import com.badlogic.gdx.Gdx
+import com.badlogic.gdx.InputMultiplexer
+import com.badlogic.gdx.Preferences
 import com.badlogic.gdx.Screen
 import com.badlogic.gdx.math.GridPoint2
 import com.badlogic.gdx.math.Vector2
+import ktx.preferences.get
 
 class GameScreen : Screen, IPhysics {
     private val tileWorld = TileWorld(64, TutorialLevel())
-    private val world: World = GameWorld.create()
+    private val world: World = GameWorld().create()
     private val playerId: Int
     private val playerEntity: Entity
     private val tagManager: TagManager = world.getSystem(TagManager::class.java)
     private val worldRenderer = GameRenderer(world, tileWorld, RenderSettings(16, 20f, 11.5f))
-    private val eventManager: EventManager = world.getRegistered(EventManager::class.java)
+    private val eventManager: EventManager = world.getRegistered("eventManager")
+    private val prefs: Preferences = world.getRegistered("preferences")
+    private val multiplexer: InputMultiplexer = InputMultiplexer()
 
     init {
         val playerArchetype = getPlayerArchetype(ArchetypeBuilder(), world)
         playerId = world.create(playerArchetype)
         playerEntity = initializePlayer(playerId, world, tagManager)
+        multiplexer.addProcessor(GameInput(eventManager, prefs["inputSettings", InputSettings()]))
+        Gdx.input.inputProcessor = multiplexer
     }
 
     private fun initializePlayer(playerId: Int, world: World, tagManager: TagManager): Entity {
