--- core/src/main/kotlin/com/abysl/gdxcrawler/settings/PreferenceManager.kt
@@ -1,3 +1,0 @@
-package com.abysl.gdxcrawler.settings
-
-class PreferenceManager
