+++ core/src/main/kotlin/com/abysl/gdxcrawler/settings/SettingNames.kt
@@ -1,0 +1,5 @@
+package com.abysl.gdxcrawler.settings
+
+enum class SettingNames {
+    INPUT_SETTINGS
+}
