--- core/src/main/kotlin/com/abysl/gdxcrawler/settings/Settings.kt
@@ -1,6 +1,0 @@
-package com.abysl.gdxcrawler.settings
-
-data class Settings(
-    val inputSettings: InputSettings,
-    val renderSettings: RenderSettings,
-)
